requirements.txt
crewai==1.6.1
gradio==6.0.2
google-generativeai==0.8.5
langchain==0.3.27
langchain-google-genai==0.3.11
litellm==1.80.5
requests==2.32.5
Pillow==12.0.0
pydantic==2.12.4
python-dotenv==1.2.1
PyYAML==6.0.3